Change from compress(inplace=bool) to compress() vs. compression()
Rationale: Doing inplace compression or leaving self unchanged is a
significant difference in control flow. We do not want to hide this in
an optional argument, or change it accidentally via **kwargs.

This is also the reason to hide compress_svd() and compress_var() from
the user: compress_var() must be renamed to compression_var(). For
consistency, we would also want to add compression_svd() and
compress_var(). Avoid complexity and just go with compress() and
compression().

Also rename sweep_sites to var_sites. The parameter has nothing to do
with sweeping (=the motion of going from left to right and back).

Fixes will follow.