compare strings with '==' instead of 'is'
Whether comparing strings with 'is' returns the expected result depends
on the context and on the Python implementation.  'is' checks whether
the left and right hand sides are the same objects, but we can well have
two different string objects which represent the same string.  E.g.
CPython 3.5.1 has the following behaviour:

$ python
Python 3.5.1 ...
Type "help", "copyright", "credits" or "license" for more information.
>>> x = 'abc'
>>> x is 'abc'
True
>>> y = b'abc'.decode('ascii')
>>> x is y
False
>>> x == y
True
>>>